Resolve circular imports in native default controls
