+++ packages/native/vite.config.ts
@@ -14,7 +14,7 @@ export default defineConfig(({ command }) => ({
       formats: ['cjs', 'es'],
     },
     rollupOptions: {
-      external: ['react', 'react-native', 'react-querybuilder'],
+      external: ['react', 'react-native', 'react-querybuilder', '@react-native-picker/picker'],
     },
     sourcemap: true,
     target: 'es2020',
